// frontend/src/components/newapi/ratio/types.ts
export type RatioOptions = {
  ModelPrice: string;
  ModelRatio: string;
  CacheRatio: string;
  CreateCacheRatio: string;
  CompletionRatio: string;
  ImageRatio: string;
  AudioRatio: string;
  AudioCompletionRatio: string;
  ExposeRatioEnabled: boolean;

  GroupRatio: string;
  UserUsableGroups: string;
  GroupGroupRatio: string;
  "group_ratio_setting.group_special_usable_group": string;
  AutoGroups: string;
  DefaultUseAutoGroup: boolean;
};